api/pyproject.toml
[project]
name = "workout-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.133.1",
    "pydantic[standard]>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.37",
]
